requirements.txt
Flask[async]
gunicorn
python-dotenv
openai
langgraph
langchain-core
langchain-openai
asyncpg
ipython
langchain_community